Move staking to its own folder
